Adjust the route params on Home screen
